refactor(analytics-react-native): extract framework-free ConnectivityChecker
Move the Android connectivity logic (currentConnectivity, hasInternetCapability,
network-callback registration, and same-state dedupe) out of
AmplitudeReactNativeConnectivityModule into a new framework-free
ConnectivityChecker that depends only on android.net.* and a listener lambda.
The module is now a thin React bridge that delegates to it and keeps only the
React surface (promise resolution, RCTDeviceEventEmitter emit, and the
hasActiveCatalystInstance() guard).

Behavior-preserving — including the backlog #5 desync where the deduped state
advances before the module decides whether to drop the emit. Mirrors
Amplitude-Kotlin's AndroidNetworkConnectivityChecker and makes the connectivity
logic unit-testable with Robolectric + mockk without React on the classpath.
