+++ packages/analytics-react-native/android/src/main/java/com/amplitude/reactnative/AmplitudeReactNativeConnectivityModule.kt
@@ -1,11 +1,5 @@
 package com.amplitude.reactnative
 
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.Network
-import android.net.NetworkCapabilities
-import android.net.NetworkRequest
-import android.os.Build
 import com.facebook.react.bridge.Arguments
 import com.facebook.react.bridge.Promise
 import com.facebook.react.bridge.ReactApplicationContext
@@ -24,33 +18,29 @@ private const val CONNECTIVITY_EVENT_NAME = "AmplitudeNetworkConnectivityChanged
  * request/response [AmplitudeReactNativeModule] so its callback lifecycle stays
  * isolated.
  *
- * Connectivity is monitored with [ConnectivityManager]. `registerDefaultNetworkCallback`
- * is API 24+, while `minSdkVersion` is 21, so we fall back to
- * `registerNetworkCallback(NetworkRequest, callback)` on API 21–23.
+ * The Android networking logic lives in the framework-free [ConnectivityChecker]
+ * (so it can be unit-tested without React Native on the classpath); this module
+ * is the thin React bridge around it — it seeds JS via [getNetworkConnectivityStatus]
+ * and forwards connectivity changes as [CONNECTIVITY_EVENT_NAME] events.
  */
 @ReactModule(name = CONNECTIVITY_MODULE_NAME)
 class AmplitudeReactNativeConnectivityModule(
     private val reactContext: ReactApplicationContext
 ) : ReactContextBaseJavaModule(reactContext) {
 
-    private val connectivityManager: ConnectivityManager?
-        get() = reactContext.getSystemService(Context.CONNECTIVITY_SERVICE) as? ConnectivityManager
-
-    private var networkCallback: ConnectivityManager.NetworkCallback? = null
-
-    // Last known connectivity state, used to seed JS and to dedupe.
-    @Volatile
-    private var isConnected = true
+    private val connectivityChecker = ConnectivityChecker(reactContext) { connected ->
+        emitConnectivityChange(connected)
+    }
 
     override fun getName(): String {
         return CONNECTIVITY_MODULE_NAME
     }
 
     // Required so `NativeEventEmitter` doesn't warn on the JS side. The actual
-    // registration is lazy (see [registerNetworkCallback]).
+    // registration is lazy (see [ConnectivityChecker.start]).
     @ReactMethod
     fun addListener(eventName: String) {
-        registerNetworkCallback()
+        connectivityChecker.start()
     }
 
     @ReactMethod
@@ -66,69 +56,7 @@ class AmplitudeReactNativeConnectivityModule(
      */
     @ReactMethod
     fun getNetworkConnectivityStatus(promise: Promise) {
-        promise.resolve(Arguments.createMap().apply { putBoolean("isConnected", currentConnectivity()) })
-    }
-
-    @Synchronized
-    private fun registerNetworkCallback() {
-        if (networkCallback != null) {
-            return
-        }
-        val manager = connectivityManager ?: return
-
-        // Seed the current state so we don't wait for the first change to know
-        // whether we're online.
-        isConnected = currentConnectivity()
-
-        val callback = object : ConnectivityManager.NetworkCallback() {
-            override fun onAvailable(network: Network) {
-                handleConnectivityChange(true)
-            }
-
-            override fun onLost(network: Network) {
-                handleConnectivityChange(false)
-            }
-
-            override fun onCapabilitiesChanged(network: Network, capabilities: NetworkCapabilities) {
-                handleConnectivityChange(hasInternetCapability(capabilities))
-            }
-        }
-
-        try {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                manager.registerDefaultNetworkCallback(callback)
-            } else {
-                val request = NetworkRequest.Builder()
-                    .addCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
-                    .build()
-                manager.registerNetworkCallback(request, callback)
-            }
-            networkCallback = callback
-        } catch (e: SecurityException) {
-            // Some devices throw if ACCESS_NETWORK_STATE is missing; leave the
-            // callback unregistered and rely on the seeded state.
-        }
-    }
-
-    @Synchronized
-    private fun unregisterNetworkCallback() {
-        val callback = networkCallback ?: return
-        try {
-            connectivityManager?.unregisterNetworkCallback(callback)
-        } catch (e: IllegalArgumentException) {
-            // Callback was not registered; ignore.
-        }
-        networkCallback = null
-    }
-
-    private fun handleConnectivityChange(connected: Boolean) {
-        // Dedupe repeated same-state callbacks (e.g. onCapabilitiesChanged firing
-        // multiple times) so JS doesn't flush-storm.
-        if (isConnected == connected) {
-            return
-        }
-        isConnected = connected
-        emitConnectivityChange(connected)
+        promise.resolve(Arguments.createMap().apply { putBoolean("isConnected", connectivityChecker.currentConnectivity()) })
     }
 
     private fun emitConnectivityChange(connected: Boolean) {
@@ -141,39 +69,14 @@ class AmplitudeReactNativeConnectivityModule(
             .emit(CONNECTIVITY_EVENT_NAME, params)
     }
 
-    private fun currentConnectivity(): Boolean {
-        val manager = connectivityManager ?: return true
-        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            val network = manager.activeNetwork ?: return false
-            val capabilities = manager.getNetworkCapabilities(network) ?: return false
-            hasInternetCapability(capabilities)
-        } else {
-            @Suppress("DEPRECATION")
-            manager.activeNetworkInfo?.isConnected ?: false
-        }
-    }
-
-    // NET_CAPABILITY_VALIDATED was added in API 23 (M); on API 21-22 it is never
-    // reported, so requiring it there would make the device read as permanently
-    // offline. Gate the validation requirement behind M and require only
-    // INTERNET on 21-22.
-    private fun hasInternetCapability(capabilities: NetworkCapabilities): Boolean {
-        val hasInternet = capabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
-        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            hasInternet && capabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_VALIDATED)
-        } else {
-            hasInternet
-        }
-    }
-
     override fun invalidate() {
-        unregisterNetworkCallback()
+        connectivityChecker.stop()
         super.invalidate()
     }
 
     @Deprecated("Deprecated in Java")
     override fun onCatalystInstanceDestroy() {
-        unregisterNetworkCallback()
+        connectivityChecker.stop()
         @Suppress("DEPRECATION")
         super.onCatalystInstanceDestroy()
     }
